Fix warning

+++ physicsEngine/src/utils/pool/FixedSizePool.inl
@@ -16,7 +16,7 @@ template<class BaseType> FixedSizePool<BaseType>::FixedSizePool(const std::strin
 
     //initialize pool: each element contains address of next element and last one contains 0
     unsigned char* p = pool;
-    int count = this->maxElements;
+    int count = (int)this->maxElements;
     while (--count) {
         *(void**)p = (p + this->maxElementSize);
         p += this->maxElementSize;
